+++ infra/terraform/data.tf
@@ -26,30 +26,31 @@ resource "random_password" "db" {
 }
 
 resource "aws_db_instance" "main" {
-  identifier                   = local.name
-  engine                       = "postgres"
-  engine_version               = "16"
-  instance_class               = var.db_instance_class
-  allocated_storage            = 50
-  max_allocated_storage        = 500
-  storage_type                 = "gp3"
-  storage_encrypted            = true
-  db_name                      = "agarha"
-  username                     = "agarha_owner"
-  password                     = random_password.db.result
-  multi_az                     = var.db_multi_az
-  db_subnet_group_name         = module.vpc.database_subnet_group_name
-  vpc_security_group_ids       = [aws_security_group.data.id]
-  parameter_group_name         = aws_db_parameter_group.pg.name
-  backup_retention_period      = local.prod ? 14 : 3
-  backup_window                = "01:00-02:00"
-  maintenance_window           = "Fri:02:30-Fri:03:30"
-  deletion_protection          = local.prod
-  skip_final_snapshot          = !local.prod
-  final_snapshot_identifier    = local.prod ? "${local.name}-final" : null
-  performance_insights_enabled = true
-  auto_minor_version_upgrade   = true
-  copy_tags_to_snapshot        = true
+  identifier                      = local.name
+  engine                          = "postgres"
+  engine_version                  = "16"
+  instance_class                  = var.db_instance_class
+  allocated_storage               = 50
+  max_allocated_storage           = 500
+  storage_type                    = "gp3"
+  storage_encrypted               = true
+  db_name                         = "agarha"
+  username                        = "agarha_owner"
+  password                        = random_password.db.result
+  multi_az                        = var.db_multi_az
+  db_subnet_group_name            = module.vpc.database_subnet_group_name
+  vpc_security_group_ids          = [aws_security_group.data.id]
+  parameter_group_name            = aws_db_parameter_group.pg.name
+  backup_retention_period         = local.prod ? 14 : 3
+  backup_window                   = "01:00-02:00"
+  maintenance_window              = "Fri:02:30-Fri:03:30"
+  deletion_protection             = local.prod
+  skip_final_snapshot             = !local.prod
+  final_snapshot_identifier       = local.prod ? "${local.name}-final" : null
+  performance_insights_enabled    = true
+  enabled_cloudwatch_logs_exports = ["postgresql", "upgrade"]
+  auto_minor_version_upgrade      = true
+  copy_tags_to_snapshot           = true
 }
 
 resource "aws_elasticache_subnet_group" "redis" {